requirements.txt
pykrx>=1.0.45
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.1
python-dotenv>=1.0.0
streamlit>=1.35.0
finance-datareader>=0.9.50
